Category components updated

+++ src/components/Category.js
@@ -1,18 +1,23 @@
 import React, { useEffect, useState} from 'react';
-import { Container, Table, Form, FormControl } from 'react-bootstrap';
+import { Button, Container, Table } from 'react-bootstrap';
 import {Link} from "react-router-dom";
-import axios from 'axios';
-import CategoryAdd from './CategoryAdd';
 import CategoryApi from './CategoryApi';
-
+import swal from 'sweetalert';
+import Pagination from 'react-js-pagination';
 
 const Category = () => {
     const [categories, setCategories] = useState(null);
+    const [currentPage, setCurrentPage] = useState(0);
+    const [perPage, setPerPage] = useState(0);
+    const [total, setTotal] = useState(0);
 
-    const fetchCategories = () => {
-        CategoryApi.getCategories().then(res => {
+    const fetchCategories = (currentPage) => {
+        CategoryApi.getCategories(currentPage).then(res => {
             console.log(res.data);
-            setCategories(res.data);
+            setCategories(res.data.data);
+            setPerPage(res.data.per_page);
+            setCurrentPage(res.data.current_page);
+            setTotal(res.data.total);
         })
         .catch(err => {
             console.log(err);
@@ -20,7 +25,7 @@ const Category = () => {
     }
 
     useEffect(() => {
-        fetchCategories();
+        fetchCategories(0);
     }, [])
 
     function renderCategories(){
@@ -46,26 +51,45 @@ const Category = () => {
 
         console.log("Hay categorías");
         console.log(categories);
+
         return categories.map((category => (
             <tr key={category.id}>
                 <th>{category.name}</th>
                 <td>{category.posts_count}</td>
                 <td>
-                    <Link className="btn btn-warning"
+                    <Link
                     to={'/categories/edit/'+category.id}>
-                        Editar
+                        <Button variant="secondary">
+                            Editar
+                        </Button>
                     </Link>
                 </td>
                 <td>
                     <button 
                         className="btn btn-danger"
                         onClick={() => {
-                            CategoryApi.deleteCategory(category.id)
-                            .then(fetchCategories)
-                            .catch(err => {
-                                alert('Error al borrar el post');
-                            });
-                        }}
+                            swal({
+                                title: "¿Está seguro?",
+                                text: (category.posts_count > 1) 
+                                      ? "Borrar la categoría borrará los " + category.posts_count + " posts relacionados."
+                                      : (category.posts_count) == 1
+                                          ? "Borrar la categoría borrará 1 post relacionado."
+                                          : "Borrará permanentemente la categoría",
+                                icon: "warning",
+                                buttons: true,
+                                dangerMode: true,
+                            })
+                            .then((willDelete) => {
+                                if(willDelete){
+                                    CategoryApi.deleteCategory(category.id)
+                                    .then(fetchCategories)
+                                    .catch(err => {
+                                        console.log(err);
+                                        swal("Error", "Hubo un error al eliminar la categoría", "error");
+                                    })
+                                }
+                            })}
+                        }
                     >
                         Borrar
                     </button>
@@ -76,33 +100,40 @@ const Category = () => {
 
     return (
         <Container>
-        <div className="d-flex justify-content-center my-4">
-          <div className="col-xs-12 col-sm-11 col-md-11 col-lg-10 col-xl-10">
-            <header>
-              <h5 className="mt-3 mb-3">Categorías</h5>
-            </header>
-            {/* Create category button */}
-            <Link to="/categories/add">
-                <a className="btn btn-secondary btn-lg btn-block p-1" role="button">Crear categoría</a>
-            </Link>
-            <table className="table">
-                <thead>
-                    <tr>
-                    <th scope="col">Nombre</th>
-                    <th scope="col">Cantidad de posts</th>
-                    <th scope="col">Editar</th>
-                    <th scope="col">Borrar</th>
-                    </tr>
-                </thead>
-                <tbody>
-                  {renderCategories()}
-                </tbody>
-            </table>
-            {/* Pagination buttons */}
-            <div className="d-flex justify-content-center mb-4">
+            <div className="m-4">
+                <h5 className="mt-3 mb-3">Categorías</h5>
+
+                <Link to="/categories/add">
+                    <Button variant="primary" className="mb-3">
+                        Crear categoría
+                    </Button>
+                </Link>
+                
+                <Table striped bordered hover responsive>
+                    <thead>
+                        <tr>
+                        <th scope="col">Nombre</th>
+                        <th scope="col">Cantidad de posts</th>
+                        <th scope="col">Editar</th>
+                        <th scope="col">Borrar</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                    {renderCategories()}
+                    </tbody>
+                </Table>
             </div>
-          </div>
-        </div>
+
+            <div className="m-4">
+                <Pagination
+                    activePage = { currentPage }
+                    totalItemsCount = { total }
+                    itemsCountPerPage = { perPage }
+                    onChange={(pageNumber) => fetchCategories(pageNumber) }
+                    itemClass = "page-item"
+                    linkClass = "page-link"
+                />
+            </div>
       </Container>
     )
 }
